feat: update route response document
