+++ TODO.md
@@ -7,3 +7,5 @@
 - [ ] Future-231126.5: 泛型路由注册
 - [ ] Future-231126.6: 查询参数考虑是否要支持数组;
 - [ ] Future-231126.7: 查询参数值不允许为map;
+- [ ] Future-231203.8: 模型不支持嵌入;
+- [ ] Future-231203.9: POST/PATCH/PUT方法最多支持2个结构体参数
