--- container.go
@@ -1,6 +1,0 @@
-package fastapi
-
-// Container 一个路由实例
-type Container struct {
-	kind int // 1: 泛型路由，2：路由组
-}
