+++ openapi/swagger.go
@@ -2,7 +2,12 @@ package openapi
 
 import (
 	"github.com/Chendemo12/fastapi-tool/helper"
+	"github.com/Chendemo12/fastapi/pathschema"
+	"github.com/Chendemo12/fastapi/utils"
 	"net/http"
+	"reflect"
+	"strconv"
+	"strings"
 )
 
 // Contact 联系方式, 显示在 info 字段内部
@@ -108,6 +113,25 @@ type Parameter struct {
 	Schema  *ParameterSchema `json:"schema,omitempty" description:"字段模型"`
 }
 
+func (p *Parameter) FromQModel(model *QModel) *Parameter {
+	p.Name = model.JsonName()
+	p.Description = model.SchemaDesc()
+	p.Required = model.IsRequired()
+	p.Default = GetDefaultV(model.Tag, model.SchemaType())
+	p.Schema = &ParameterSchema{
+		Type:  model.SchemaType(),
+		Title: model.SchemaTitle(),
+	}
+
+	if model.InPath {
+		p.In = InPath
+	} else {
+		p.In = InQuery
+	}
+
+	return p
+}
+
 type ModelContentSchema interface {
 	Schema() map[string]any
 	SchemaType() DataType
@@ -188,6 +212,50 @@ func (o *Operation) MarshalJSON() ([]byte, error) {
 	}
 
 	return helper.JsonMarshal(orm)
+}
+
+// RequestBodyFrom 从 *openapi.BaseModelMeta 转换成 openapi 的请求体 RequestBody
+func (o *Operation) RequestBodyFrom(model *BaseModelMeta) *Operation {
+	o.RequestBody = &RequestBody{}
+	if model != nil {
+		o.RequestBody.Required = model.IsRequired()
+		o.RequestBody.Content = &PathModelContent{
+			MIMEType: MIMEApplicationJSON,
+			Schema:   model,
+		}
+	}
+
+	return o
+}
+
+// ResponseFrom 从 *openapi.BaseModelMeta 转换成 openapi 的响应实例
+func (o *Operation) ResponseFrom(model *BaseModelMeta) *Operation {
+	if model == nil { // 若返回值为空，则设置为空
+		model = &BaseModelMeta{}
+	}
+
+	m := make([]*Response, 2) // 200 + 422
+	// 200 接口处注册的返回值
+	m[0] = &Response{
+		StatusCode:  http.StatusOK,
+		Description: http.StatusText(http.StatusOK),
+		Content: &PathModelContent{
+			MIMEType: MIMEApplicationJSON,
+			Schema:   model,
+		},
+	}
+	// 422 所有接口默认携带的请求体校验错误返回值
+	m[1] = &Response{
+		StatusCode:  http.StatusUnprocessableEntity,
+		Description: http.StatusText(http.StatusUnprocessableEntity),
+		Content: &PathModelContent{
+			MIMEType: MIMEApplicationJSON,
+			Schema:   &ValidationError{},
+		},
+	}
+
+	o.Responses = m
+	return o
 }
 
 // PathItem 路由选项，由于同一个路由可以存在不同的操作方法，因此此选项可以存在多个 Operation
@@ -279,11 +347,9 @@ func (o *OpenApi) AddDefinition(meta SchemaIface) *OpenApi {
 	return o
 }
 
-// AddPathItem 手动添加路由对象, 不存在则新建, 存在则更新
-func (o *OpenApi) AddPathItem(path string) *PathItem {
-	// 修改路径格式为fastapi路径格式
-	// 主要区别在于用{}标识路径参数,而非:
-
+// 查询路由对象, 不存在则新建
+func (o *OpenApi) getPath(path string) *PathItem {
+	// 修改路径格式为FastApi路径格式, 主要区别在于用"{}"标识路径参数,而非":"
 	path = ToFastApiRoutePath(path) // 修改路径格式
 
 	for _, item := range o.Paths.Paths {
@@ -314,6 +380,7 @@ func (o *OpenApi) modelFrom(swagger *RouteSwagger) {
 		for _, inner := range swagger.RequestModel.InnerSchema() {
 			o.AddDefinition(inner)
 		}
+		// 处理数组类型
 		if swagger.RequestModel.itemModel != nil {
 			o.AddDefinition(swagger.RequestModel.itemModel)
 			for _, inner := range swagger.RequestModel.itemModel.InnerSchema() {
@@ -327,6 +394,7 @@ func (o *OpenApi) modelFrom(swagger *RouteSwagger) {
 		for _, inner := range swagger.ResponseModel.InnerSchema() {
 			o.AddDefinition(inner)
 		}
+		// 处理数组类型
 		if swagger.ResponseModel.itemModel != nil {
 			o.AddDefinition(swagger.ResponseModel.itemModel)
 			for _, inner := range swagger.ResponseModel.itemModel.InnerSchema() {
@@ -338,12 +406,13 @@ func (o *OpenApi) modelFrom(swagger *RouteSwagger) {
 
 func (o *OpenApi) pathFrom(swagger *RouteSwagger) {
 	// 存在相同路径，不同方法的路由选项
-	item := o.AddPathItem(swagger.Url)
+	item := o.getPath(swagger.Url)
 
 	// 构造路径参数
 	pathParams := make([]*Parameter, len(swagger.PathFields))
 	for no, q := range swagger.PathFields {
-		p := QModelToParameter(q)
+		p := &Parameter{}
+		p.FromQModel(q)
 		p.Deprecated = swagger.Deprecated
 
 		pathParams[no] = p
@@ -352,7 +421,8 @@ func (o *OpenApi) pathFrom(swagger *RouteSwagger) {
 	// 构造查询参数
 	queryParams := make([]*Parameter, len(swagger.QueryFields))
 	for no, q := range swagger.QueryFields {
-		p := QModelToParameter(q)
+		p := &Parameter{}
+		p.FromQModel(q)
 		p.Deprecated = swagger.Deprecated
 		queryParams[no] = p
 	}
@@ -363,10 +433,15 @@ func (o *OpenApi) pathFrom(swagger *RouteSwagger) {
 		Description: swagger.Description,
 		Tags:        swagger.Tags,
 		Parameters:  append(pathParams, queryParams...),
-		RequestBody: MakeOperationRequestBody(swagger.RequestModel),
-		Responses:   MakeOperationResponses(swagger.ResponseModel),
 		Deprecated:  swagger.Deprecated,
 	}
+	if utils.Has[string]([]string{http.MethodGet, http.MethodDelete}, swagger.Method) {
+		// GET/DELETE 无请求体，不显示
+		operation.RequestBody = nil
+	} else {
+		operation.RequestBodyFrom(swagger.RequestModel)
+	}
+	operation.ResponseFrom(swagger.ResponseModel)
 
 	// 绑定到操作方法
 	switch swagger.Method {
@@ -415,4 +490,111 @@ func (o *OpenApi) Schema() []byte {
 	}
 
 	return o.cache
+}
+
+// ToFastApiRoutePath 将 fiber.App 格式的路径转换成 FastApi 格式的路径
+//
+//	Example:
+//	必选路径参数：
+//		Input: "/api/rcst/:no"
+//		Output: "/api/rcst/{no}"
+//	可选路径参数：
+//		Input: "/api/rcst/:no?"
+//		Output: "/api/rcst/{no}"
+//	常规路径：
+//		Input: "/api/rcst/no"
+//		Output: "/api/rcst/no"
+func ToFastApiRoutePath(path string) string {
+	paths := strings.Split(path, pathschema.PathSeparator) // 路径字符
+	// 查找路径中的参数
+	for i := 0; i < len(paths); i++ {
+		if strings.HasPrefix(paths[i], pathschema.PathParamPrefix) {
+			// 识别到路径参数
+			if strings.HasSuffix(paths[i], pathschema.OptionalQueryParamPrefix) {
+				// 可选路径参数
+				paths[i] = "{" + paths[i][1:len(paths[i])-1] + "}"
+			} else {
+				paths[i] = "{" + paths[i][1:] + "}"
+			}
+		}
+	}
+
+	return strings.Join(paths, pathschema.PathSeparator)
+}
+
+// ReflectKindToType 转换reflect.Kind为swagger类型说明
+//
+//	@param	ReflectKind	reflect.Kind	反射类型,不进一步对指针类型进行上浮
+func ReflectKindToType(kind reflect.Kind) (name DataType) {
+	switch kind {
+
+	case reflect.Array, reflect.Slice, reflect.Chan:
+		name = ArrayType
+	case reflect.String:
+		name = StringType
+	case reflect.Bool:
+		name = BoolType
+	default:
+		if reflect.Bool < kind && kind <= reflect.Uint64 {
+			name = IntegerType
+		} else if reflect.Float32 <= kind && kind <= reflect.Complex128 {
+			name = NumberType
+		} else {
+			name = ObjectType
+		}
+	}
+
+	return
+}
+
+// IsFieldRequired 从tag中判断此字段是否是必须的
+func IsFieldRequired(tag reflect.StructTag) bool {
+	for _, name := range []string{GinValidateTagName, DefaultValidateTagName} {
+		bindings := strings.Split(utils.QueryFieldTag(tag, name, ""), ",") // binding 存在多个值
+		for i := 0; i < len(bindings); i++ {
+			if strings.TrimSpace(bindings[i]) == DefaultParamRequiredLabel {
+				return true
+			}
+		}
+	}
+
+	return false
+}
+
+// GetDefaultV 从Tag中提取字段默认值
+func GetDefaultV(tag reflect.StructTag, otype DataType) (v any) {
+	defaultV := utils.QueryFieldTag(tag, DefaultValueTagNam, "")
+
+	if defaultV == "" {
+		v = nil
+	} else { // 存在默认值
+		switch otype {
+
+		case StringType:
+			v = defaultV
+		case IntegerType:
+			v, _ = strconv.Atoi(defaultV)
+		case NumberType:
+			v, _ = strconv.ParseFloat(defaultV, 64)
+		case BoolType:
+			v, _ = strconv.ParseBool(defaultV)
+		default:
+			v = defaultV
+		}
+	}
+	return
+}
+
+func assignModelNames(field reflect.StructField, fatherModel reflect.Type) (string, string) {
+	var pkg, name string
+	if utils.IsAnonymousStruct(field.Type) {
+		// 未命名的结构体类型, 没有名称, 分配包名和名称
+		name = fatherModel.Name() + "Model"
+		pkg = fatherModel.PkgPath() + AnonymousModelNameConnector + name
+	} else { // 命名的结构体
+		pkg = field.Type.String() // 关联模型
+		name = field.Type.Name()
+	}
+
+	return pkg, name
 }
